#146 multipl test for controllers

+++ backend/vurrukkulluk/src/main/java/com/lydia/vurrukkulluk/controller/ImageController.java
@@ -10,6 +10,7 @@ import com.lydia.vurrukkulluk.service.RecipeService;
 import com.lydia.vurrukkulluk.service.UserService;
 import com.lydia.vurrukkulluk.util.SecurityUtil;
 import com.lydia.vurrukkulluk.util.UserImageUtil;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -31,7 +32,7 @@ import java.util.Objects;
 // GET: In Postman, make a HTTP GET request to
 // http://localhost:8080/image/{id} You will see the image in the response body downloaded from the database.
 
-
+@AllArgsConstructor
 @RestController
 @RequestMapping("api/image")
 @CrossOrigin
@@ -58,42 +59,82 @@ public class ImageController {
   @ResponseStatus(value = HttpStatus.OK)
   @PostMapping()
   public String add(@RequestParam("image") MultipartFile file, @RequestParam String type, @RequestParam int id) throws IOException {
-    Image image = new Image(file);
-    image = imageService.saveImage(image);
+
     switch (type) {
       case "user" -> {
         if (!securityUtil.isAuthorizedUserOrAdmin(id)) {return "not authorized";}
+        Image image = new Image(file);
+        image = imageService.saveImage(image);
         User user = userService.getUserById(id);
         userService.setImageInUser(user,image);
       }
       case "recipe" -> {
         Recipe recipe = recipeService.getRecipeById(id);
         if (!securityUtil.isAuthorizedUserOrAdmin(recipe.getUser().getId())){return "not authorized";}
+        Image image = new Image(file);
+        image = imageService.saveImage(image);
         recipeService.setImageInRecipe(recipe,image);
       }
       case "article" -> {
-//        if (!securityUtil.isAdmin()) {return "not authorized";}
+        if (!securityUtil.isAdmin()) {return "not authorized";}
+        Image image = new Image(file);
+        image = imageService.saveImage(image);
         articleService.setImageInArticle(id,image);
 
       }
     }
 
-    return "Saved image with id: " + id;
+    return "Saved image";
   }
 
   @PutMapping("/{id}")
-  public String put(@RequestParam("image") MultipartFile file, @PathVariable int id) throws IOException{
+  public String put(@RequestParam("image") MultipartFile file, @PathVariable int imageId, @PathVariable String type,@PathVariable int objectId) throws IOException{
 
-    Image image = new Image(id,file);
-    imageService.updateImage(image);
-    return "Updated image with id:" + id;
+    String authorizationMessage = imageAuthorization(type,objectId);
+    if (!authorizationMessage.equals("authorized")) {
+      return authorizationMessage;
+    }
+
+    imageService.updateImage(imageId,file);
+    return "Updated image with id:" + imageId;
   }
 
   @DeleteMapping("/{id}")
-  public String delete(@PathVariable int id) {
+  public String delete(@PathVariable int id, @PathVariable String type, @PathVariable int objectId ) {
+
+    String authorizationMessage = imageAuthorization(type,objectId);
+    if (!authorizationMessage.equals("authorized")) {
+      return authorizationMessage;
+    }
+
     imageService.deleteImage(id);
     return "Deleted image with id: " + id;
   }
+  private String imageAuthorization(String type, int objectId){
+    switch (type) {
+      case "user" -> {
+        if (!securityUtil.isAuthorizedUserOrAdmin(objectId)) {
+          return "not authorized";
+        }
+      }
+      case "recipe" -> {
+        Recipe recipe = recipeService.getRecipeById(objectId);
+        if (!securityUtil.isAuthorizedUserOrAdmin(recipe.getUser().getId())) {
+          return "not authorized";
+        }
+      }
+      case "article" -> {
+        if (!securityUtil.isAdmin()) {
+          return "not authorized";
+        }
+      }
+      default -> {
+        return "invalid type";
+      }
+    }
+    return "authorized";
+    }
+
 
 }
